MyFastAPIs/requirements.txt
pymongo[srv]>=4.6.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0